+++ src/cli/middleware.js
@@ -3,6 +3,7 @@ import { cleanFlag, translateArgs } from '../helpers/args.js'
 import { translateCommand } from '../helpers/cmds.js'
 import { detectVoltaPin, getCurrentPackageManager } from '../helpers/get.js'
 import { getPackageConfiguration } from '../packages/list.js'
+import { setPackageVersion } from '../helpers/set.js'
 
 const middleware = async (yargs) => {
   const pkg = {
@@ -20,6 +21,7 @@ const middleware = async (yargs) => {
     cleanFlag(yargs, '--use')
     cleanFlag(yargs, '-u')
     pkg.cmd = yargs.use
+    await setPackageVersion(yargs.$0, yargs.use)
   }
 
   if ('pin' in yargs) {