+++ vitest.config.ts
@@ -1,16 +1,28 @@
-import { defineConfig } from "vitest/config";
+import { configDefaults, defineConfig } from 'vitest/config'
+
+const include = [
+  'src/**/*',
+]
+
+const exclude = [
+  ...configDefaults.exclude,
+  'src/alias/*',
+  'test{,s}/**',
+  'test{,-*}.{js,cjs,mjs,ts,tsx,jsx}',
+  '**/*{.,-}test.{js,cjs,mjs,ts,tsx,jsx}',
+]
 
 export default defineConfig({
   test: {
     globals: true,
-    reporters: ["verbose"],
+    reporters: ['verbose'],
     coverage: {
       all: true,
       reporter: ["text", "html", "lcov"],
       statements: 22,
-      branches: 50,
-      functions: 26,
-      lines: 23,
+      branches: 48,
+      functions: 27,
+      lines: 22,
     },
   },
 });